requirements.txt
flask==3.0.0
flask-cors==4.0.0
googletrans==4.0.0-rc1
deep-translator==1.11.4
google-generativeai==0.3.2
pysrt==1.1.2
webvtt-py==0.4.6
python-dotenv==1.0.0
requests==2.31.0
chardet==5.2.0
pydantic==2.5.0
aiohttp==3.9.1
asyncio==3.4.3
gunicorn==21.2.0
